+++ src/core/injections/controller/checks.py
@@ -24,6 +24,7 @@ import string
 import base64
 import traceback
 from collections import OrderedDict 
+from src.core.convert import hexdecode
 from src.utils import menu
 from src.thirdparty.six.moves import input as _input
 from src.thirdparty.six.moves import urllib as _urllib
@@ -524,6 +525,9 @@ def check_http_s(url):
           else:
             url = "http://" + url
         settings.SCHEME = (_urllib.parse.urlparse(url).scheme.lower() or "http") if not menu.options.force_ssl else "https"
+        if menu.options.force_ssl and settings.VERBOSITY_LEVEL >= 1:
+          debug_msg = "Forcing usage of SSL/HTTPS requests."
+          print(settings.print_debug_msg(debug_msg))
       else:
         err_msg = "Invalid target URL has been given." 
         print(settings.print_critical_msg(err_msg))
@@ -1013,12 +1017,12 @@ def recognise_payload(payload):
       decoded_payload = base64.b64decode(payload)
       if re.match(settings.HEX_RECOGNITION_REGEX, payload):
         settings.MULTI_ENCODED_PAYLOAD.append("hexencode")
-        decoded_payload = decoded_payload.decode("hex")
+        decoded_payload = hexdecode(decoded_payload)
 
   elif re.match(settings.HEX_RECOGNITION_REGEX, payload):
     is_decoded = True
     settings.MULTI_ENCODED_PAYLOAD.append("hexencode")
-    decoded_payload = payload.decode("hex")
+    decoded_payload = hexdecode(payload)
     if (len(payload) % 4 == 0) and \
       re.match(settings.BASE64_RECOGNITION_REGEX, decoded_payload) and \
       not re.match(settings.HEX_RECOGNITION_REGEX, decoded_payload):