+++ src/core/requests/redirection.py
@@ -46,7 +46,7 @@ def do_check(url):
         print(settings.print_warning_msg(warn_msg))
         return Request(redirected_url, 
                            headers = newheaders,
-                           origin_req_host = req.get_origin_req_host(), 
+                           # origin_req_host = req.get_origin_req_host(), 
                            unverifiable = True
                            ) 
       else: 
@@ -63,7 +63,7 @@ def do_check(url):
       newheaders = dict((k,v) for k,v in req.headers.items() if k.lower() not in ("content-length", "content-type"))
       return self.parent.open(_urllib.request.Request(req.get_full_url(), 
                               headers = newheaders, 
-                              origin_req_host = req.get_origin_req_host(), 
+                              # origin_req_host = req.get_origin_req_host(), 
                               unverifiable = True)
                               )
 
@@ -93,7 +93,7 @@ def do_check(url):
         try:
           url = menu.options.url
           try:
-            response = _urllib.request.urlopen(url)
+            response = _urllib.request.urlopen(url, timeout=settings.TIMEOUT)
           except _urllib.error.HTTPError as e:
             try:
               authline = e.headers.get('www-authenticate', '')  
@@ -106,7 +106,7 @@ def do_check(url):
               authhandler.add_password(realm, url, username, password)
               opener = _urllib.request.build_opener(authhandler)
               _urllib.request.install_opener(opener)
-              result = _urllib.request.urlopen(url)
+              result = _urllib.request.urlopen(url, timeout=settings.TIMEOUT)
             except AttributeError:
               pass
         except _urllib.error.HTTPError as e: